Review changes

+++ util/storage-samples/storage-owner-sample/src/main/java/org/fidoalliance/fdo/storage/OwnerSysModule.java
@@ -19,7 +19,7 @@ public class OwnerSysModule implements Module {
   private Composite state;
   private final DataSource dataSource;
   private static final int MAX_READ = 1277;
-  private static final int CBOR_TRUE = 0xF5;
+  private static final byte CBOR_TRUE = (byte) 0xF5;
 
 
   /**
@@ -146,7 +146,7 @@ public class OwnerSysModule implements Module {
       case FdoSys.KEY_ACTIVE: {
 
         byte[] content = new OwnerDbManager().getSystemResourceContent(dataSource, resId);
-        if (content.length == 1 && content[0] == ((byte) CBOR_TRUE)) {
+        if (content.length == 1 && content[0] == CBOR_TRUE) {
           message.set(Const.SECOND_KEY, true);
         } else {
           message.set(Const.SECOND_KEY, false);
